+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add a formatter option for space within return list parenthesis.
 - Add error recovery for statements.
 - Add support for parsing empty statements.
+- Add an inspection for missing script symbols.
 
 ### Changed
 - The build window will now display internal compiler errors as well.